ai-recruitment-be/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv

langchain
langchain-chroma
langchain-huggingface
langchain-openai